+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+- Does not report screen events if there's no title ([#32](https://github.com/PostHog/posthog-android/pull/32))
+
 ## 3.0.0-alpha.4 - 2023-10-03
 
 - Fixed small issues after code review